+++ src/datastructpy/node.py
@@ -1,0 +1,17 @@
+class Node:
+    """
+    A class representing a node in a binary search tree (BST).
+
+    Attributes
+    ----------
+    key : int
+        The value stored in the node.
+    left : Node, optional
+        The left child node (default is None).
+    right : Node, optional
+        The right child node (default is None).
+    """
+    def __init__(self, key):
+        self.key = key
+        self.left = None
+        self.right = None
