+++ src/datastructpy/non-linear/trees/binary_search_tree.py
@@ -1,21 +1,3 @@
-class Node:
-    """
-    A class representing a node in a binary search tree (BST).
-
-    Attributes
-    ----------
-    key : int
-        The value stored in the node.
-    left : Node, optional
-        The left child node (default is None).
-    right : Node, optional
-        The right child node (default is None).
-    """
-    def __init__(self, key):
-        self.key = key
-        self.left = None
-        self.right = None
-
 class BinarySearchTree:
     """
     A class representing a binary search tree (BST).
@@ -163,4 +145,3 @@ class BinarySearchTree:
         print(bst.root.left.key)  # Output: 5 (left child of root)
         print(bst.root.right.key) # Output: 15 (right child of root)
         """
-
