+++ tests/non-linear/trees/binary_search_tree/test_delete.py
@@ -1,0 +1,2 @@
+from datastructpy.non_linear.trees.binary_search_trees import BinarySearchTree
+
