+++ tests/non-linear/trees/binary_search_tree/test_search.py
@@ -1,0 +1,1 @@
+from datastructpy.non_linear.trees.binary_search_trees import BinarySearchTree
