+++ tests/test_datastructpy.py
@@ -1,1 +1,3 @@
 from datastructpy import datastructpy
+
+# To do: Delete file later
